requirements.txt
google-generativeai
gtts